+++ web/src/nav/GlobalNavbar.tsx
@@ -27,6 +27,9 @@ import { FiltersToTypeAndValue } from '../../../shared/src/search/interactive/ut
 import { SearchModeToggle } from '../search/input/interactive/SearchModeToggle'
 import { Link } from '../../../shared/src/components/Link'
 import { convertPlainTextToInteractiveQuery } from '../search/input/helpers'
+import { VersionContextDropdown } from './VersionContextDropdown'
+import { VersionContextProps } from '../../../shared/src/search/util'
+import { VersionContext } from '../schema/site.schema'
 
 interface Props
     extends SettingsCascadeProps,
@@ -40,7 +43,8 @@ interface Props
         PatternTypeProps,
         CaseSensitivityProps,
         InteractiveSearchProps,
-        SmartSearchFieldProps {
+        SmartSearchFieldProps,
+        VersionContextProps {
     history: H.History
     location: H.Location<{ query: string }>
     authenticatedUser: GQL.IUser | null
@@ -65,6 +69,8 @@ interface Props
     splitSearchModes: boolean
     interactiveSearchMode: boolean
     toggleSearchMode: (event: React.MouseEvent<HTMLAnchorElement>) => void
+    setVersionContext: (versionContext: string) => void
+    availableVersionContexts: VersionContext[] | undefined
 
     /** For testing only. Used because reactstrap's Popover is incompatible with react-test-renderer. */
     hideNavLinks: boolean
@@ -194,6 +200,7 @@ export class GlobalNavbar extends React.PureComponent<Props, State> {
                                     navbarSearchState={this.props.navbarSearchQueryState}
                                     onNavbarQueryChange={this.props.onNavbarQueryChange}
                                     lowProfile={!this.props.isSearchRelatedPage}
+                                    versionContext={this.props.versionContext}
                                 />
                             )
                         ) : (
@@ -207,6 +214,15 @@ export class GlobalNavbar extends React.PureComponent<Props, State> {
                                                 interactiveSearchMode={this.props.interactiveSearchMode}
                                             />
                                         )}
+                                        <VersionContextDropdown
+                                            history={this.props.history}
+                                            navbarSearchQuery={this.props.navbarSearchQueryState.query}
+                                            caseSensitive={this.props.caseSensitive}
+                                            patternType={this.props.patternType}
+                                            versionContext={this.props.versionContext}
+                                            setVersionContext={this.props.setVersionContext}
+                                            availableVersionContexts={this.props.availableVersionContexts}
+                                        />
                                         <SearchNavbarItem
                                             {...this.props}
                                             navbarSearchState={this.props.navbarSearchQueryState}