+++ web/src/repo/RepoRevContainer.tsx
@@ -38,6 +38,7 @@ import { RepoSettingsAreaRoute } from './settings/RepoSettingsArea'
 import { RepoSettingsSideBarItem } from './settings/RepoSettingsSidebar'
 import { ErrorMessage } from '../components/alerts'
 import * as H from 'history'
+import { VersionContextProps } from '../../../shared/src/search/util'
 
 /** Props passed to sub-routes of {@link RepoRevContainer}. */
 export interface RepoRevContainerContext
@@ -53,7 +54,8 @@ export interface RepoRevContainerContext
             Exclude<keyof RepoContainerContext, 'onDidUpdateRepository' | 'onDidUpdateExternalLinks'>
         >,
         PatternTypeProps,
-        CaseSensitivityProps {
+        CaseSensitivityProps,
+        VersionContextProps {
     repo: GQL.IRepository
     rev: string
     resolvedRev: ResolvedRev
@@ -75,7 +77,8 @@ interface RepoRevContainerProps
         ThemeProps,
         ActivationProps,
         PatternTypeProps,
-        CaseSensitivityProps {
+        CaseSensitivityProps,
+        VersionContextProps {
     routes: readonly RepoRevContainerRoute[]
     repoSettingsAreaRoutes: readonly RepoSettingsAreaRoute[]
     repoSettingsSidebarItems: readonly RepoSettingsSideBarItem[]
@@ -232,6 +235,7 @@ export class RepoRevContainer extends React.PureComponent<RepoRevContainerProps,
             setCaseSensitivity: this.props.setCaseSensitivity,
             repoSettingsAreaRoutes: this.props.repoSettingsAreaRoutes,
             repoSettingsSidebarItems: this.props.repoSettingsSidebarItems,
+            versionContext: this.props.versionContext,
         }
 
         return (