+++ web/src/search/input/SearchNavbarItem.tsx
@@ -9,6 +9,7 @@ import { LazyMonacoQueryInput } from './LazyMonacoQueryInput'
 import { QueryInput } from './QueryInput'
 import { ThemeProps } from '../../../../shared/src/theme'
 import { SettingsCascadeProps } from '../../../../shared/src/settings/settings'
+import { VersionContextProps } from '../../../../shared/src/search/util'
 import { KEYBOARD_SHORTCUT_FOCUS_SEARCHBAR } from '../../keyboardShortcuts/keyboardShortcuts'
 
 interface Props
@@ -17,7 +18,8 @@ interface Props
         CaseSensitivityProps,
         SmartSearchFieldProps,
         SettingsCascadeProps,
-        ThemeProps {
+        ThemeProps,
+        VersionContextProps {
     location: H.Location
     history: H.History
     navbarSearchState: QueryState