+++ web/src/search/input/toggles/Toggles.tsx
@@ -11,12 +11,14 @@ import { submitSearch } from '../../helpers'
 import { QueryInputToggle } from './QueryInputToggle'
 import { isErrorLike } from '../../../../../shared/src/util/errors'
 import CodeBracketsIcon from 'mdi-react/CodeBracketsIcon'
+import { VersionContextProps } from '../../../../../shared/src/search/util'
 
 export interface TogglesProps
     extends PatternTypeProps,
         CaseSensitivityProps,
         SettingsCascadeProps,
-        Partial<Pick<InteractiveSearchProps, 'filtersInQuery'>> {
+        Partial<Pick<InteractiveSearchProps, 'filtersInQuery'>>,
+        VersionContextProps {
     navbarSearchQuery: string
     history: H.History
     location: H.Location
@@ -34,7 +36,7 @@ export const Toggles: React.FunctionComponent<TogglesProps> = (props: TogglesPro
         window.context.experimentalFeatures.structuralSearch === 'disabled'
 
     const submitOnToggle = (args: { newPatternType: SearchPatternType } | { newCaseSensitivity: boolean }): void => {
-        const { history, navbarSearchQuery, filtersInQuery } = props
+        const { history, navbarSearchQuery, filtersInQuery, versionContext } = props
         const searchQueryNotEmpty = navbarSearchQuery !== '' || (filtersInQuery && !isEmpty(filtersInQuery))
         const shouldSubmitSearch = props.hasGlobalQueryBehavior && searchQueryNotEmpty
         const activation = undefined
@@ -51,6 +53,7 @@ export const Toggles: React.FunctionComponent<TogglesProps> = (props: TogglesPro
                 source,
                 patternType,
                 caseSensitive,
+                versionContext,
                 activation,
                 filtersInQuery,
             })