+++ web/src/views/QueryInputInViewContent.tsx
@@ -6,8 +6,9 @@ import { SearchButton } from '../search/input/SearchButton'
 import { SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { QueryState, submitSearch } from '../search/helpers'
 import H from 'history'
+import { VersionContextProps } from '../../../shared/src/search/util'
 
-interface Props extends SettingsCascadeProps, PatternTypeProps, CaseSensitivityProps {
+interface Props extends SettingsCascadeProps, PatternTypeProps, CaseSensitivityProps, VersionContextProps {
     implicitQueryPrefix: string
 
     location: H.Location