+++ web/src/views/ViewContent.tsx
@@ -9,12 +9,17 @@ import { CaseSensitivityProps, PatternTypeProps } from '../search'
 import { SettingsCascadeProps } from '../../../shared/src/settings/settings'
 import { hasProperty } from '../../../shared/src/util/types'
 import { isObject } from 'lodash'
+import { VersionContextProps } from '../../../shared/src/search/util'
 import { ChartViewContent } from './ChartViewContent'
 
 const isMarkupContent = (input: unknown): input is MarkupContent =>
     isObject(input) && hasProperty('value')(input) && typeof input.value === 'string'
 
-export interface ViewContentProps extends SettingsCascadeProps, PatternTypeProps, CaseSensitivityProps {
+export interface ViewContentProps
+    extends SettingsCascadeProps,
+        PatternTypeProps,
+        CaseSensitivityProps,
+        VersionContextProps {
     viewContent: View['content']
     location: H.Location
     history: H.History